+++ spath/src/spec/segment.rs
@@ -14,7 +14,7 @@
 
 use std::fmt;
 
-use crate::spec::functions::FunctionRegistry;
+use crate::spec::function::FunctionRegistry;
 use crate::spec::query::Queryable;
 use crate::spec::select_wildcard;
 use crate::spec::selector::Selector;
