+++ spath/src/spec/selector/filter.rs
@@ -21,9 +21,9 @@ use super::name::Name;
 use super::Selector;
 use crate::node::LocatedNode;
 use crate::path::NormalizedPath;
-use crate::spec::functions::FunctionExpr;
-use crate::spec::functions::FunctionRegistry;
-use crate::spec::functions::SPathValue;
+use crate::spec::function::FunctionExpr;
+use crate::spec::function::FunctionRegistry;
+use crate::spec::function::SPathValue;
 use crate::spec::query::Query;
 use crate::spec::query::QueryKind;
 use crate::spec::query::Queryable;
