+++ spath/src/spec/selector/index.rs
@@ -18,7 +18,7 @@ use std::fmt;
 
 use num_traits::ToPrimitive;
 
-use crate::spec::functions::FunctionRegistry;
+use crate::spec::function::FunctionRegistry;
 use crate::spec::query::Queryable;
 use crate::ConcreteVariantArray;
 use crate::LocatedNode;
